+++ service/internal/images/images.go
@@ -54,7 +54,7 @@ func New(c Config) (*Images, error) {
 	return i, nil
 }
 
-func (i *Images) AMI(ctx context.Context, obj interface{}) (string, error) {
+func (i *Images) AMI(ctx context.Context, obj interface{}, flatcarReleaseVersion string) (string, error) {
 	cr, err := meta.Accessor(obj)
 	if err != nil {
 		return "", microerror.Mask(err)
@@ -70,7 +70,7 @@ func (i *Images) AMI(ctx context.Context, obj interface{}) (string, error) {
 		return "", microerror.Mask(err)
 	}
 
-	ami, err := key.AMI(key.Region(cl), re)
+	ami, err := key.AMI(key.Region(cl), re, flatcarReleaseVersion)
 	if err != nil {
 		return "", microerror.Mask(err)
 	}